refactor(native): split runtime and macOS engine modules
