+++ native/src/engine/cef_mac/scheme.h
@@ -1,0 +1,16 @@
+#ifndef PROTON_ENGINE_CEF_MAC_SCHEME_H
+#define PROTON_ENGINE_CEF_MAC_SCHEME_H
+
+#include "include/capi/cef_frame_capi.h"
+#include "include/capi/cef_request_capi.h"
+#include "include/capi/cef_resource_handler_capi.h"
+#include "include/capi/cef_scheme_capi.h"
+
+cef_resource_handler_t *CEF_CALLBACK proton_engine_scheme_create(
+    cef_scheme_handler_factory_t *self,
+    cef_browser_t *browser,
+    cef_frame_t *frame,
+    const cef_string_t *scheme_name,
+    cef_request_t *request);
+
+#endif
